Chore MemberVO와 PaymentVO폴더 안으로 위치 변경
