+++ src/main/java/dao/Payment/PaymentDaoImpl.java
@@ -1,6 +1,6 @@
 package dao.Payment;
 
-import domain.PaymentVO;
+import domain.Payment.PaymentVO;
 import java.sql.*;
 import java.util.*;
